Adjustments for motion correction and bad alignment
The second step of motion correction was inducing horizontal and vertical shifts in datasets in which alignment was not possible or bad due to low image quality, absence of clear landmarks, or large drifts in the field of view.
Fixed by taking the second motion correction out of the pipeline (manual alignment is already semi-automatic and will not induce more human error) and also by including a function to check for bad alignment sessions. In this case, the algorithm will exclude this sessions from the pipeline for cell extraction and replace the excluded session by the session with the best average correlation with all the other sessions.