Select ROI from the FOV
Changes to include a function to select an ROI from the FOV. Since V4 miniscopes have a larger FOV and a longer working distance, the borders of the lens can be within the FOV. Optical noise on the borders may compromise motion correction and cell detection, especially when the miniscope was not properlly fixed during recordings. This update includes a function to select ROIs from the FOV of each session separatelly before motion correction, or after concatenation of the motion corrected videos and before cell detection. This has the potential to improve motion correction and reduce flase positive cell detection, which seems to be the case in our preliminary analysis with a small dataset.